fix(push/pull): set URL before pushing/pulling
This is needed, since `git clone` sets the remote URL to the absolute
path. This means that the GitHub build is broken, or also the tutorial2
build would refer to the "tutorial" folder for that, which is broken.

This fix still leaves the path in the reflog of the
"another-downstream". A better solution would be to clone this only when
needed, but a first try proved more complicated than expected... Since
this is only harming reproducibility a little bet, and probably nobody
ever looks at the reflog of that hidden repo, this should be fine.

Fixes: https://github.com/miallo/nuggit/issues/46